Convert PemToKeyPairTask to PemToKeyPairObservable
Fixes #1364.

At code level it's reusing the same class, but with RxJava's Observable should do things better than AsyncTasks anyway.
